fix: eliminate all 70 type-check-ts errors (tsc --noEmit clean)
Critical fixes:
- L2PSHashService: access account.nonce directly instead of stale details.content.nonce path
- Sync: remove extra l2psJoinedUids argument silently dropped by exchangeL2PSParticipation
- handleIdentityRequest: restore SDK imports now that abstraction barrel exports ZK types

High-priority fixes:
- Add ffjavascript ambient type declarations for ZK wrappers (BunSnarkjs, BunPlonk)
- Import ValidityData type in L2PSHashService
- Add BunRequest type extending Request with params in bunServer
- Fix getCurveFromName second arg from object to boolean in BunPlonkWrapper

Test/harness fixes:
- Add `as Uint8Array` casts for SDK getIdentity() union type (~30 sites)
- Fix HDNodeWallet/Wallet type in gcr_identity_xm_smoke
- Add missing normalizeTokenPointerEntry import in token_shared
- Add missing methodNames in token_script_complex_policy_escrow_state_machine
- Fix fetch wrapper type cast, Set<string> generic, union narrowing in framework
- Fix MetricsCollector port label type, fhe_test variadic args, proof-verifier imports
- Remove unused @ts-expect-error directives
- Fix rateLimiter.test mock spread args

Bump demosdk to ^2.11.4 (abstraction barrel now exports ZK identity types).

+++ better_testing/loadgen/src/features/consensus/consensus_rollback_smoke.ts
@@ -33,11 +33,11 @@ export async function runConsensusRollbackSmoke() {
   await demos.connectWallet(wallets[0]!, { algorithm: "ed25519" })
 
   const { publicKey } = await demos.crypto.getIdentity("ed25519")
-  const senderAddress = uint8ArrayToHex(publicKey)
+  const senderAddress = uint8ArrayToHex(publicKey as Uint8Array)
   const recipientDemos = new Demos()
   await recipientDemos.connect(bootstrap)
   await recipientDemos.connectWallet(wallets[1]!, { algorithm: "ed25519" })
-  const recipientAddress = uint8ArrayToHex((await recipientDemos.crypto.getIdentity("ed25519")).publicKey)
+  const recipientAddress = uint8ArrayToHex((await recipientDemos.crypto.getIdentity("ed25519")).publicKey as Uint8Array)
 
   const senderNonceBefore = await getAddressNonceViaRpc(bootstrap, senderAddress, "consensus:rollback:nonce:before")
   if (typeof senderNonceBefore !== "number") {
